requirements.txt
amqp==2.6.1
asgiref==3.7.2
billiard==3.6.4.0
celery==4.4.7
celery-slack==0.4.1
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==4.0.0
django-extensions==3.2.1
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
ephem==3.7.7.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
sentry-sdk==1.24.0
six==1.16.0
sqlparse==0.4.4
timeago==1.0.16
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==1.26.16
vine==1.3.0
whitenoise==6.4.0
yarg==0.1.9
gunicorn==20.1.0
gevent